+++ components/OAuth2.0/Authorization.c
@@ -1,6 +1,6 @@
 #include <stdio.h>
 #include <stdlib.h>
-#include "authorization.h"
+#include "Authorization.h"
 #include "OauthHttpLocalServer.h"
 #include "SpiffsManger.h"
 #include "OauthAPICall.h"
