+++ components/OAuth2.0/CMakeLists.txt
@@ -1,4 +1,4 @@
-idf_component_register(SRCS "authorization.c" "OauthAPICall.c" "OauthHttpLocalServer.c"
+idf_component_register(SRCS "Authorization.c" "OauthAPICall.c" "OauthHttpLocalServer.c"
                     INCLUDE_DIRS "include"
                     
                     REQUIRES esp_psram mdns heap esp_event esp-tls
