+++ components/OAuth2.0/OauthAPICall.c
@@ -1,5 +1,5 @@
 #include "OauthAPICall.h"
-#include "authorization.h"
+#include "Authorization.h"
 
 static const char *TAG = "HTTP";
 APIBuffer_t *OAuthBuffer;
