+++ components/OAuth2.0/include/OauthHttpLocalServer.h
@@ -5,7 +5,7 @@ extern "C"
 #ifndef OAUTH_HTTP_SERVER
 #define OAUTH_HTTP_SERVER
 
-#include "authorization.h"
+#include "Authorization.h"
 
 extern QueueHandle_t SendCodeFromHttpToTask;
